new idea for tcp message sending
I found out: it is impossible to send template classes through signal/slot in qt.
But: why not set up a TcpMessage class, that consists of a QVector of  QVariant (or union) types that can be whatever is needed. Also one QVector or simply one int that holds the information about what type of message it is. For example: i2cUpdateMessage ->  the data in the vector are int, float, float, int, bool... in this order.