fixed bugs with setting rate settings. Changed to tabs
Now there are tabs for settings, status informations and so on
inside the used to be empty space. Have to think about the design (I like it like this but maybe we don't need the sliders to be visible all the time). Going to implement more ubx settings and status information soon.