The RatePlot shows now ... something. Added version 1.0.1 of the daemon.
A few bugs are still there:
The overlap between existing points and non-existing points is not detected at all. This results in some weird shapes.
The static cast from quint64 (msecs since program start) to qreal (should be float on armhf) seems to be faulty.
There is a segmentation fault occuring only when you close the gui.