Automatically save login data for ftp upload: Working!
The file will automatically be generated when providing the login data via "--lftp-login" flag. Using this flag will prompt you to type in the login data.  After doing this once the data is saved encrypted (but could technically be decrypted if you know the macaddress of the system it was encrypted on). It will automatically decrypt the login data and it should upload ifiles to the ftp server once a day.