added a modified version of gui for android
Can compile and pack the gui but it crashes on start. In theory it should be possible to create a slightly modified version of the desktop gui for android devices. Next step would be running the apk on a simulated androd phone in android studio and using a debugger to see what breaks.